+++ Backend/internal/services/resume_review_helpers.go
@@ -1,0 +1,364 @@
+package services
+
+import (
+	"encoding/json"
+	"errors"
+	"fmt"
+	"strings"
+
+	"Backend/internal/models"
+)
+
+func fallbackResumeReview(blocks []models.ResumeTextBlock) (*models.ResumeReview, []models.ResumeReviewItem) {
+	score := 70
+	summary := "主要セクションを確認しました。具体性の強化が改善ポイントです。"
+	items := make([]models.ResumeReviewItem, 0)
+	bbox, _ := json.Marshal([]float64{20, 20, 260, 80})
+	items = append(items, models.ResumeReviewItem{
+		PageNumber: 1,
+		BBox:       string(bbox),
+		Severity:   "info",
+		Message:    "内容は整理されていますが、成果の具体性や背景の説明が不足しがちです。",
+		Suggestion: "成果を数値で示し、役割や工夫点・課題を一文ずつ補足してください。",
+	})
+	return &models.ResumeReview{
+		Score:   score,
+		Summary: summary,
+	}, items
+}
+
+func fallbackResumeReviewDetailed(blocks []models.ResumeTextBlock) (*models.ResumeReview, []models.ResumeReviewItem) {
+	score := 70
+	summary := "内容を確認しました。各項目の具体性を高めると説得力が増します。"
+	items := buildHeuristicItems(blocks, 8)
+	if len(items) == 0 {
+		return fallbackResumeReview(blocks)
+	}
+	return &models.ResumeReview{
+		Score:   score,
+		Summary: summary,
+	}, items
+}
+
+func buildResumeText(blocks []models.ResumeTextBlock, maxLen int) string {
+	if len(blocks) == 0 {
+		return ""
+	}
+	var b strings.Builder
+	for _, block := range blocks {
+		line := strings.TrimSpace(block.Text)
+		if line == "" {
+			continue
+		}
+		b.WriteString(fmt.Sprintf("[P%dB%d] %s\n", block.PageNumber, block.BlockIndex, line))
+		if b.Len() >= maxLen {
+			break
+		}
+	}
+	return b.String()
+}
+
+func decodeJSON(raw string, out any) error {
+	raw = strings.TrimSpace(raw)
+	if raw == "" {
+		return errors.New("empty response")
+	}
+	if err := json.Unmarshal([]byte(raw), out); err == nil {
+		return nil
+	}
+	start := strings.Index(raw, "{")
+	end := strings.LastIndex(raw, "}")
+	if start >= 0 && end > start {
+		return json.Unmarshal([]byte(raw[start:end+1]), out)
+	}
+	return errors.New("invalid JSON response")
+}
+
+func mapReviewItems(blocks []models.ResumeTextBlock, aiItems []aiReviewItem) []models.ResumeReviewItem {
+	if len(aiItems) == 0 {
+		return nil
+	}
+	result := make([]models.ResumeReviewItem, 0, len(aiItems))
+	for _, item := range aiItems {
+		if strings.TrimSpace(item.Quote) == "" {
+			continue
+		}
+		var block *models.ResumeTextBlock
+		foundByIndex := false
+		if item.PageHint > 0 && item.BlockIndex > 0 {
+			block = findBlockByIndex(blocks, item.PageHint, item.BlockIndex)
+			if block != nil {
+				foundByIndex = true
+			}
+		}
+		if block == nil && runeLen(item.Quote) >= 6 {
+			block = findBestBlock(blocks, item.Quote, item.PageHint)
+		}
+		if block == nil {
+			continue
+		}
+		if !foundByIndex && !quoteInBlock(item.Quote, block.Text) {
+			continue
+		}
+		severity := strings.ToLower(item.Severity)
+		if severity == "" {
+			severity = "info"
+		}
+		result = append(result, models.ResumeReviewItem{
+			PageNumber: block.PageNumber,
+			BBox:       block.BBox,
+			Severity:   severity,
+			Message:    item.Message,
+			Suggestion: item.Suggestion,
+		})
+	}
+	return result
+}
+
+func findBestBlock(blocks []models.ResumeTextBlock, quote string, pageHint int) *models.ResumeTextBlock {
+	quoteNorm := normalizeText(quote)
+	if quoteNorm == "" {
+		return nil
+	}
+
+	var best *models.ResumeTextBlock
+	bestScore := 0
+	for i := range blocks {
+		block := &blocks[i]
+		if pageHint > 0 && block.PageNumber != pageHint {
+			continue
+		}
+		blockNorm := normalizeText(block.Text)
+		if blockNorm == "" {
+			continue
+		}
+		score := textMatchScore(blockNorm, quoteNorm)
+		if score > bestScore {
+			bestScore = score
+			best = block
+		}
+	}
+
+	if best != nil {
+		return best
+	}
+
+	for i := range blocks {
+		block := &blocks[i]
+		blockNorm := normalizeText(block.Text)
+		score := textMatchScore(blockNorm, quoteNorm)
+		if score > bestScore {
+			bestScore = score
+			best = block
+		}
+	}
+	return best
+}
+
+func findBlockByIndex(blocks []models.ResumeTextBlock, pageHint int, blockIndex int) *models.ResumeTextBlock {
+	for i := range blocks {
+		block := &blocks[i]
+		if block.PageNumber == pageHint && block.BlockIndex == blockIndex {
+			return block
+		}
+	}
+	return nil
+}
+
+func normalizeText(s string) string {
+	s = strings.ToLower(strings.TrimSpace(s))
+	s = strings.ReplaceAll(s, " ", "")
+	s = strings.ReplaceAll(s, "　", "")
+	return s
+}
+
+func quoteInBlock(quote string, blockText string) bool {
+	return strings.Contains(normalizeText(blockText), normalizeText(quote))
+}
+
+func runeLen(s string) int {
+	return len([]rune(strings.TrimSpace(s)))
+}
+
+func buildHeuristicItems(blocks []models.ResumeTextBlock, max int) []models.ResumeReviewItem {
+	if len(blocks) == 0 || max <= 0 {
+		return nil
+	}
+	result := make([]models.ResumeReviewItem, 0, max)
+	seen := make(map[string]bool)
+	for _, block := range blocks {
+		text := strings.TrimSpace(block.Text)
+		if runeLen(text) < 12 {
+			continue
+		}
+		label, message, suggestion := classifyBlock(text)
+		if label == "" {
+			continue
+		}
+		key := fmt.Sprintf("%d-%d-%s", block.PageNumber, block.BlockIndex, label)
+		if seen[key] {
+			continue
+		}
+		seen[key] = true
+		result = append(result, models.ResumeReviewItem{
+			PageNumber: block.PageNumber,
+			BBox:       block.BBox,
+			Severity:   "info",
+			Message:    message,
+			Suggestion: suggestion,
+		})
+		if len(result) >= max {
+			return result
+		}
+	}
+	if len(result) == 0 {
+		for _, block := range blocks {
+			text := strings.TrimSpace(block.Text)
+			if runeLen(text) < 16 {
+				continue
+			}
+			result = append(result, models.ResumeReviewItem{
+				PageNumber: block.PageNumber,
+				BBox:       block.BBox,
+				Severity:   "info",
+				Message:    "この記述は成果や役割の具体性が読み取りづらいです。",
+				Suggestion: "成果の数値や担当範囲、工夫点を一文ずつ補足してください。",
+			})
+			if len(result) >= max {
+				return result
+			}
+		}
+	}
+	return result
+}
+
+func classifyBlock(text string) (string, string, string) {
+	switch {
+	case strings.Contains(text, "志望") || strings.Contains(text, "動機"):
+		return "motivation",
+			"志望動機の根拠が抽象的に見えます。",
+			"企業の事業や職種と自分の経験の接点を1文で明示し、具体的な業務貢献を追記してください。"
+	case strings.Contains(text, "自己PR") || strings.Contains(text, "自己ＰＲ"):
+		return "pr",
+			"自己PRが強みの列挙にとどまっています。",
+			"成果の数値、工夫した点、再現性が分かる行動を1文ずつ追加してください。"
+	case strings.Contains(text, "学歴"):
+		return "", "", ""
+	case strings.Contains(text, "職歴"):
+		return "", "", ""
+	case strings.Contains(text, "資格") || strings.Contains(text, "免許"):
+		return "license",
+			"資格が応募職種にどう活かせるかが伝わりづらいです。",
+			"資格で得たスキルと職務での活用例を一文追加してください。"
+	case strings.Contains(text, "得意") || strings.Contains(text, "特技") || strings.Contains(text, "スキル"):
+		return "skill",
+			"スキルの記載が抽象的で実務イメージが湧きにくいです。",
+			"使用期間、具体的な成果物、担当範囲を補足してください。"
+	case strings.Contains(text, "学生時代"):
+		return "student",
+			"活動の規模や成果が読み取りづらいです。",
+			"人数・期間・結果などの具体的な数値を補足してください。"
+	}
+	return "generic",
+		"この記述は成果や役割の具体性が読み取りづらいです。",
+		"成果の数値、役割、工夫点を一文ずつ補足してください。"
+}
+
+func selectReviewBlocks(blocks []models.ResumeTextBlock, max int) []models.ResumeTextBlock {
+	if len(blocks) == 0 || max <= 0 {
+		return nil
+	}
+	selected := make([]models.ResumeTextBlock, 0, max)
+	seenPages := make(map[int]bool)
+	for _, block := range blocks {
+		text := strings.TrimSpace(block.Text)
+		if runeLen(text) < 12 {
+			continue
+		}
+		if strings.HasSuffix(text, "：") || strings.HasSuffix(text, ":") {
+			continue
+		}
+		selected = append(selected, block)
+		seenPages[block.PageNumber] = true
+		if len(selected) >= max {
+			return selected
+		}
+	}
+	if len(selected) < max {
+		for _, block := range blocks {
+			if seenPages[block.PageNumber] {
+				continue
+			}
+			text := strings.TrimSpace(block.Text)
+			if runeLen(text) < 8 {
+				continue
+			}
+			selected = append(selected, block)
+			if len(selected) >= max {
+				break
+			}
+		}
+	}
+	return selected
+}
+
+func buildBlockList(blocks []models.ResumeTextBlock) string {
+	if len(blocks) == 0 {
+		return ""
+	}
+	var b strings.Builder
+	for _, block := range blocks {
+		line := strings.TrimSpace(block.Text)
+		if line == "" {
+			continue
+		}
+		b.WriteString(fmt.Sprintf("[P%dB%d] %s\n", block.PageNumber, block.BlockIndex, line))
+	}
+	return b.String()
+}
+
+func textMatchScore(block, quote string) int {
+	if block == "" || quote == "" {
+		return 0
+	}
+	if strings.Contains(block, quote) || strings.Contains(quote, block) {
+		return 100
+	}
+	blockTokens := splitTokens(block)
+	quoteTokens := splitTokens(quote)
+	if len(blockTokens) == 0 || len(quoteTokens) == 0 {
+		return 0
+	}
+	score := 0
+	for _, token := range quoteTokens {
+		for _, b := range blockTokens {
+			if token == b {
+				score += 10
+				break
+			}
+		}
+	}
+	return score
+}
+
+func splitTokens(s string) []string {
+	s = strings.ReplaceAll(s, "。", " ")
+	s = strings.ReplaceAll(s, "、", " ")
+	s = strings.ReplaceAll(s, "\n", " ")
+	parts := strings.Fields(s)
+	if len(parts) == 0 {
+		return nil
+	}
+	seen := make(map[string]bool)
+	result := make([]string, 0, len(parts))
+	for _, part := range parts {
+		if len(part) < 2 {
+			continue
+		}
+		if !seen[part] {
+			seen[part] = true
+			result = append(result, part)
+		}
+	}
+	return result
+}
